Begin syncing to 25.1.0 design library code.
This CL will look a bit strange, as it reverse some changes that
were already present. Unfortunately the Android auto-merger greatly
confuses our syncing tool, as it effectively creates new commits
that have the same content as old commits, and does so in a
non-linear fashion (e.g. the commit tree in 25.1.0 is not merely
built on top of 25.0.1).

This CL will get us into a state where we're ready to apply the rest
of the 25.1.0 changes in succession, and once that is completed,
we'll be at the correct (matching) state.

GitOrigin-RevId=11a895a323af06c4731e4c64a7ec2c09ac0aaf2f

PiperOrigin-RevId: 141497561