Fix password toggle taking FOCUS_DOWN events
Pressing the TAB key sends a FOCUS_DOWN event, and
now that the EditText floats in the center, the
password toggle is actually the next vertically top focusable
view.

This CL fixes it by ensuring that the EditText has the same
min-height as the password toggle, thus making the EditText
the next focusable view.

BUG: 33771077
Test: included
Change-Id: I18f5a7e419fa6ae2b13d4081e6f3e369b67ccdbe

PiperOrigin-RevId: 144259419