Fix AppBarLayout elevation jump on API 21-23
Since we use a SLA now to control elevation
changes, we've hit the issue where draw dispatching
is skipped before the Activity transition finishes
on API 21-23 devices.

This results in the SLA running, but visually you see a
jump in elevation after the activity animation has
finished (and draw commences).  To combat this, we now
jump the elevation change in layout.

Test: manual (impossible to automate it)
BUG: 29598367
Change-Id: Ia4f1c2168d363297b7e94ca9f4a0113e4f8ab271

PiperOrigin-RevId: 148950143