Fix VectorDrawable bugs in ItemActivity.
This commit fixes two bugs with VectorDrawable that were causing a crash in pre-L devices. One bug was a simple fix: using app:srcCompat instead of android:src for the FAB's icon. The other bug was more complicated, since it involved a VectorDrawable in a compound drawable. To work around the compatibility issues with this, I wrapped the VectorDrawable in a StateListDrawable and enabled AppCompat's ability to set vector drawables from resources.

PiperOrigin-RevId: 152136061