Add Gradle build for Shrine.
The Gradle build includes a workaround to make our project structure work with Gradle. Gradle does not allow any duplicate names in the project structure, so to get around this, I put the demos package under a different group.

This commit also upgrades the targetSdkVersion to 25 to match the rest of material_components and restructures the FAB click listener to fix Gradle's complaint about the String in the listener having to be final.

Gradle build for Shrine's test directory to come in a future commit.

PiperOrigin-RevId: 152527247