Fix naming for Bottom Sheet descriptions so they don't overflow on small viewports.
In order to solve overflow I have renamed the pages for BottomSheetBehavior and BottomSheetDialogFragment to "Bottom Sheets" and "Modal Bottom Sheets", reflecting their name in the spec (and usage) rather than the Android class name. This is consistent with how most other components's pages are named.

Furthermore I switched the excerpts to describe their use in the spec, and made that the first sentence in the intro for consistency.

PiperOrigin-RevId: 154995463