Make titles, paths, and headers consistent.
Follows these rules:

Title is the Component Name as it appears in spec.
Justification: Formatting and display on small width screens.

Excerpt includes the CamelCase name of the java class.
Justification: This document is not a rehash of spec, it's a specific resource for Android devs. Title and excerpt are displayed in the same context.

Header is the CamelCase name of the java class.
Justification: Similar to excerpt. This is documentation on a specific Android widget.

Path is the dash-deliminated name of the java class.
Justification: Similar to header. This is documentation on a specific Android widget.
PiperOrigin-RevId: 155134932