Remove the split base/ics/lollipop directories.
The implementation in ics/ was the only remaining implementation of that class,
so it made sense to move it directly into src/. Most of the implementations in
base/ were the only remaining implementations, and thus also made sense to move
into src/.

The remaining classes were already post-fixed with Lollipop, so I see very
little value in maintaining separate source directories for them (and I know
these separate directories have also been flagged as confusing). They've all
been moved into src/ as well now. I also renamed FloatingActionButtonLollipop
to FloatingActionButtonImplLollipop here, as that better describes what it is
a version of.

PiperOrigin-RevId: 158187999