BackLayer: Make Sibling Behavior listen to changes inside the BackLayerLayout (views changing size or relayout) to react automatically to changes in content.
Before, if there was any change to the layout inside BackLayerLayout while the back layer was already expanded, the developer had to manually call setExpanded(true) for it to animate to the new size.

Now, the behavior reacts automatically and animates the size change on its own, simplifying the code path and allowing more flexibility.

PiperOrigin-RevId: 165175310