Automated g4 rollback of changelist 165622891.
*** Reason for rollback ***

Breaks GMM

*** Original change description ***

Reposition label to adapt to box surroundings.

Move the collapsed and expanded bounds of the floating label towards where each belongs. This commit is the first stage in setting the floating label's bounds, so it avoids calculating the exact dimensions based on internal dimensions. The label positioning work is being broken up into stages since the internal dimensions of the text field will likely change either in an upcoming design rethink and while adding the dimension parameters. The final d...

***

PiperOrigin-RevId: 165664978