Add outline text field box.
This commit adds the initial outline version of the text field box. The box drawable is now created programmatically based on early versions of parameters. Parameters at this point are just hardcoded variables.

Also included: skeleton styles and attributes for the filled and outline box. These are very basic for now. Eventually, the styles will each set parameters for the box attributes, and Widget.Design.TextInputLayout.Box will set the values that each box style has in common (like height). For now, the styles pretty much just say whether the box is filled or outlined, and the setBoxAttributes() method fills in the missing information.

Not included in this commit:
- Padding
- Actual parameters set via styles/attributes
- States

PiperOrigin-RevId: 166736640