Improve backwards compatibility for box background color.
Use pre-L compatible methods for setting the background color to avoid a crash, and set the background color to default to Color.TRANSPARENT. Previously, in cases where there was no background color, the color defaulted to black in versions before 17. Now when there is no color set, the color is set to Color.TRANSPARENT instead of no color.

The Color.TRANSPARENT default color could have been restricted to pre-17, but I thought it made sense to set a transparent color for all versions, to provide a more consistent way of checking that there is no background color. I'm open to other perspectives on this.

PiperOrigin-RevId: 169300417