Make box mode attribute public.
Publicizing the box mode attribute. Box mode was previously kept internal, since it wasn't yet determined whether the box would act as mode-agnostic, and this attribute would go away, or whether there would be a concept of modes, and different behaviors to follow. This has now been decided: the box will use modes to determine functionality (although it won't rely exclusively on the mode - it will still check for null backgrounds, for example). This is the first step in supporting changing the box mode programmatically and at runtime.

PiperOrigin-RevId: 169305486