Add padding parameter and improve box creation.
This commit improves the box creation process and adds initial padding parameters for the box. For now, these values are only set on the EditText, but to truly update the padding throughout the box, the padding should be adjusted in other places (for example, the label's collapsed and expanded bounds should be updated based on the padding values). Also, some of the official values (top and bottom) don't look quite right at this stage, so they are set to values that look better for now. When padding is addressed fully, these will be set to the correct values.

This commit also adds padding back in where it would get wiped out in API 21 and 22 in ensureBackgroundDrawableStateWorkaround().

The box creation improvements prevent some unnecessary calls to box-specific methods, and a new mEditTextBoundsUpdated flag ensures that the EditText bounds are not repeatedly updated.

PiperOrigin-RevId: 170111678