Add methods for measuring collapsed text bounds.
Incorporate more robust measurement methods into CollapsingTextHelper, and use those to measure the extra top margin in TextInputLayout.

Before, there was no method to find the exact height of the text in TextInputLayout. All CollapsingTextHelper offered was a getter for the collapsed/expanded text sizes. However, this was not an accurate measurement, since the typeface could change the text dimensions. The expanded and collapsed bounds are also not reliable, because they could be set to a bigger area than the text actually takes up.

This commit adds methods to check the actual measurements of the text, rather than approximations.

PiperOrigin-RevId: 170486018