Create label cutout for outline box.
Use PorterDuff modes to create a cutout for the label when using the outline box version of TextInputLayout.

This commit focuses on creating the cutout functionality. The padding and bounds aren't the focus of this commit, and they aren't quite right yet. Measurements will be addressed in a future commit.

PiperOrigin-RevId: 170855509