Fix bug where the same temp Matrix instance was being referenced beyond its valid timeframe. There were two major bugs:
1. The temp instance toValue passed to ObjectAnimator.ofObject(view, prop, eval, toValue) does not get read right away.
Instead, ObjectAnimator stores it as a reference. This is problematic because the temp instance is used to make other calculations, and its value overwritten before ObjectAnimator reads it.

2. ImageMatrixProperty#get was returning the instance from ImageView#getImageMatrix(), which is not a copy, but the ImageView's instance itself.
This is problematic because ObjectAnimator stores it as the fromValue and reads from it on each frame. However, the ObjectAnimator is also setting new values on the same instance each frame due to animation.

The fix is to make a copy of the temp instance when it is not read right away. This way, ObjectAnimator stores a copy which is not subject to being overwritten.

PiperOrigin-RevId: 171169919