Add disabled state for outline box's stroke color.
Make the outline box respond to color state changes for the disabled state. This commit is
focused on the color changes for disabled state. The disabled state also includes a change in the box's stroke width, and that will be addressed in a future commit.

This commit introduces a color for disabled state and a field in TextInputLayout to keep track of the current box's stroke color, to prevent overwriting the box's stroke color.

PiperOrigin-RevId: 171885433