Add focused and enabled states for outline box's stroke color.
This commit makes the box's stroke color respond to the enabled and focused state. It sets the outline box's stroke to remain the default 42% black in the enabled state, and sets the user specified box stroke color on focus state.

Also included is another text field with no hint. This is to provide a way to remove focus from the text field, in order to see that the outline box only takes on the user specified color on focus.

PiperOrigin-RevId: 172353190