Add disabled color for expanded and collapsed label.
Set the expanded and collapsed label colors to the disabled color in disabled state. This commit also contains a slight refactor of the way states are handled in updateLabelState. The main thing that changed is that now, each state check does not include a check for isEnabled, since the non-enabled case is caught in what is now the first if statement, the disabled state.

PiperOrigin-RevId: 172501865