Create the baseline for box stroke width and make the stroke width changed based on state.
Add different stroke width dimensions to correspond to the different stroke widths needed for the baseline outline text field. There are two major groupings for stroke width:

1. Default width (1dp), which is used for the idle & filled, enabled, and disabled states
2. Focused width, which is used for the focused, pressed, error and hover states

These group names are not precisely definitive of the states they correspond to, but they provide the general picture and the default-focused naming is consistent with the naming used for text color in TextInputLayout.

Adding a stroke width that's larger than the one initially drawn onto the Canvas could cause the box's stroke to be cut off, since the Drawable's bounds were only allotted based on the smaller stroke. To prevent this, the outline box now creates room for half of the outline box's larger stroke in updateTextInputBoxBounds().

PiperOrigin-RevId: 172520143