Replace buttonTint and buttonTintMode with chipBackgroundColor.
Now I have a better understanding of how these attributes affect each other.
buttonTint and buttonTintMode are attributes on the View, that allow you to apply a tint on the chip drawable on top of itself.
chipBackgroundColor is an attribute on the Drawable.
It's possible to have both: think a shared chip style with white chipBackgroundColor, where the color is applied via buttonTint depending on context.

PiperOrigin-RevId: 172544568