Implement custom Chip keyboard focus support.
Left/right can move focus to the chip and close icon, after which focus jumps outside the chip.
Up/down/tab moves the focus outside the chip.

Whenever focus is newly gained, always focus the chip first. This is true even if focus is coming from the right.

We have to handle the keyboard focus explicitly, instead of relying on the ExploreByTouchHelper.
The reason is that one of our logical sub-views is the special identifier HOST_ID, which corresponds to the entire chip.
ExploreByTouchHelper seems to only work when all the sub-views are virtual views instead.

PiperOrigin-RevId: 173362171