Implement Chip ripples.
Because ChipDrawable is one drawable, the ripple is drawn over the text and icons.
It's possible to refactor this in the future so the ChipDrawable owns RippleDrawable, so that it can manage the drawing order.

Extract ripple subsystem to its own package.
Fix an issue with stateful ripple colors. Now all ripples are optimized for the pressed state, ignoring all other states.

Implement compat ripples for pre-21 Chips.

PiperOrigin-RevId: 173366665