Add dense vs. standard parameter to text field boxes.
Add style variants and dimen values to support dense and standard heights for both the filled and outline boxes, with a focus on the outline box. The filled box dimensions are also improved in this commit, but they could still use some polish.

This commit also restructures the styles a bit. Now, the Widget.MaterialComponents text field defaults to the outline box instead of the line text field. The line text field is still the default for Widget.Design's text field. Now, if someone doesn't set a style on the text field and is using the Widget.MaterialComponents theme, the text field will take on the outline style.

PiperOrigin-RevId: 173542517