Add box background color parameter to text field box.
Adds a box background parameter to the text field box. This is useful for the filled box background, and also potentially for some new developments to the outline box.

The filled box now takes on a 4% black background by default. This adds support for changing the background later on, either to support other colors, or to support other states (such as hovered, pressed, etc.)

PiperOrigin-RevId: 173600532