Add conditional source selection to MDC to deal with CoordinatorLayout move.
In 28 SDK drops, CoordinatorLayout has moved to core_ui. In preparation for
that, this CL adds select's for resources and java files related to it, as well
as an export for core_ui (temporarily).

Whenever an SDK drop occurs, that CL should change the SDK_<version>_DROP flag
to True in order to select the correct sources/resources.

PiperOrigin-RevId: 175065874