New chipIconEnabled, closeIconEnabled, and checkedIconEnabled attributes.
Before, we required you to set chipIcon, closeIcon, or checkedIcon to @null if you wanted to not show an icon.
This mainly works, but the issue appears in styles where the icon is optional, but if you want to show it, you should use a specific drawable.
Clients are then forced to look up the right drawable and set checkedIcon="@drawable/mtrl_chip_filter_checked_black".

Now, clients can do checkedIconEnabled="true".

PiperOrigin-RevId: 175071675