Nested scrolling++
This CL adds a number of overloads for the support
implementation of the nested scrolling APIs, to
add the 'type' parameter. This allow developers to
differentiate between direct touch scrolls, and
other non-direct scrolls. Primarily this is to support
dispatching the nested scrolling APIs from when a
scrolling view settles from a fling.

NestedScrollView and RecyclerView have both been updated
to dispatch these events from a fling settle. AppBarLayout
has also been updated to consume these new events.

The way the interfaces have been implemented allow us to
be API compatible without breaking anyone. Yes, the names
are ugly but so be it. If developers wish to support the
new APIs they just need to change the interface they
implement.

This CL also fixes some scroll issues in NSV and RV:
- NSV: fling not starting if scroll is completely
  consumed (since it doesn't think it is dragging).
- RV: intercept setting the wrong X/Y start values
  sometimes causing flinging in the wrong direction.

This CL includes a later fix as well
[I54655d06214751ab5533d7809e168e700fba6f96]:
Fix AppBarLayout fling + click issue

It is caused by the change in 26.0.0 where flings can
continue, even if the view is not actually moving. This
was done to allow AppBarLayout to react to "resting" flings
through nested scrolling.

The issue here is that once AppBarLayout has finished reacting
to the resting fling, and reach its end, the fling can still
continue. Therefore even though nothing is moving on screen, when
the user touches the screen they are just cancelling the fling.
Subsequent touches then do as expected (clicks, etc).

The fix here is more of a workaround than a proper fix. Here
AppBarLayout detects when it has finished its fling handling,
and then stops the nested scroll.

BUG: 22780822
Test: included for RV
Change-Id: I37de005e4fb33b5b443852058f3e1ab22ef98e20
PiperOrigin-RevId: 175155412