Add styles for bottom nav and include legacy bottom nav in demo.
Add a Widget.MaterialComponents style for bottom nav and set existing attributes to themed values. Also adds a toggle to get the legacy bottom nav, and a demo code refactor to make the code more generic. For now, the code supports just a toggle between the legacy and non-legacy bottom nav, but the newer, more generic code structure now supports easily upgrading the demo to support other versions of the bottom nav.

PiperOrigin-RevId: 176382825