Correct bottom nav code sample.
The code sample said to use the itemBackground attribute to set a color on the bottom nav, but that attribute is meant for the ripple and should not be overridden. Instead, clients should set android:background.

Also updated some color values to suggest colors that are closer to our guidance from spec.

PiperOrigin-RevId: 179946811