Create component style attribute for text input.
And assign it to the right defaults in the theme.

Since text fields is in the widget package, . I put the attr declaration in widget/../attrs.xml right on top of the theme. If text fields is ever pulled into its own package, this attribute will be moved along with the rest of TextInputLayout's attributes.

PiperOrigin-RevId: 183125979