Make the default TextInputLayout style a base style.
That way, users will have to depend on one of the box styles directly. Before, anyone that used Widget.MaterialComponents.TextInputLayout would have gotten an outline text field, which is not implied by the style name.

This commit also includes a slight reformat to make sure that style items that don't fit on one line are properly formatted.

PiperOrigin-RevId: 183130385