Use chip text in close icon's content description to clarify the icon's contentDescription.
To make it clearer what the close icon does, use the chip text inside the actual content description, and change "close" to "remove".

If the chip text is not usable (ex: it's an input chip and only spaces were entered), the screen reader will read out "chip" in place of the chip's content.

PiperOrigin-RevId: 192823693